# src/main/java/org/example/cuidadodemascotas/servicemicroservice/apis/controller/ServiceApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.example.cuidadodemascotas.servicemicroservice.apis.controller;

import org.example.cuidadodemascotas.servicemicroservice.apis.dto.ServicePageResponse;
import org.example.cuidadodemascotas.servicemicroservice.apis.dto.ServiceRequestDTO;
import org.example.cuidadodemascotas.servicemicroservice.apis.dto.ServiceResponseDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-07T01:35:32.143539700-03:00[America/Asuncion]")
@Validated
@Tag(name = "Service", description = "the Service API")
public interface ServiceApi {

    /**
     * POST /services : Crear un nuevo servicio
     *
     * @param serviceRequestDTO  (required)
     * @return Servicio creado (status code 201)
     */
    @Operation(
        operationId = "createService",
        summary = "Crear un nuevo servicio",
        tags = { "Service" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Servicio creado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ServiceResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/services",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ServiceResponseDTO> createService(
        @Parameter(name = "ServiceRequestDTO", description = "", required = true) @Valid @RequestBody ServiceRequestDTO serviceRequestDTO
    );


    /**
     * DELETE /services/{id} : Eliminar servicio (soft delete)
     *
     * @param id  (required)
     * @return Servicio eliminado (status code 204)
     */
    @Operation(
        operationId = "deleteService",
        summary = "Eliminar servicio (soft delete)",
        tags = { "Service" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Servicio eliminado")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/services/{id}"
    )
    
    ResponseEntity<Void> deleteService(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    /**
     * GET /services/{id} : Obtener servicio por ID
     *
     * @param id  (required)
     * @return Servicio obtenido (status code 200)
     */
    @Operation(
        operationId = "getServiceById",
        summary = "Obtener servicio por ID",
        tags = { "Service" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Servicio obtenido", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ServiceResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/services/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<ServiceResponseDTO> getServiceById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    /**
     * GET /services : Obtener servicios con filtros opcionales
     *
     * @param carerId  (optional)
     * @param serviceTypeId  (optional)
     * @param minPrice  (optional)
     * @param maxPrice  (optional)
     * @param page  (optional, default to 0)
     * @param size  (optional, default to 10)
     * @return Servicios obtenidos (status code 200)
     */
    @Operation(
        operationId = "getServices",
        summary = "Obtener servicios con filtros opcionales",
        tags = { "Service" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Servicios obtenidos", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ServicePageResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/services",
        produces = { "application/json" }
    )
    
    ResponseEntity<ServicePageResponse> getServices(
        @Parameter(name = "carerId", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "carerId", required = false) Long carerId,
        @Parameter(name = "serviceTypeId", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "serviceTypeId", required = false) Long serviceTypeId,
        @Parameter(name = "minPrice", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "minPrice", required = false) Double minPrice,
        @Parameter(name = "maxPrice", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "maxPrice", required = false) Double maxPrice,
        @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
        @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "10") Integer size
    );


    /**
     * GET /services/search : Buscar servicios por descripción
     *
     * @param text  (required)
     * @param page  (optional, default to 0)
     * @param size  (optional, default to 10)
     * @return Servicios encontrados (status code 200)
     */
    @Operation(
        operationId = "searchServicesByText",
        summary = "Buscar servicios por descripción",
        tags = { "Service" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Servicios encontrados", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ServicePageResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/services/search",
        produces = { "application/json" }
    )
    
    ResponseEntity<ServicePageResponse> searchServicesByText(
        @NotNull @Parameter(name = "text", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "text", required = true) String text,
        @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
        @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "10") Integer size
    );


    /**
     * PUT /services/{id} : Actualizar servicio
     *
     * @param id  (required)
     * @param serviceRequestDTO  (required)
     * @return Servicio actualizado (status code 200)
     */
    @Operation(
        operationId = "updateService",
        summary = "Actualizar servicio",
        tags = { "Service" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Servicio actualizado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ServiceResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/services/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ServiceResponseDTO> updateService(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "ServiceRequestDTO", description = "", required = true) @Valid @RequestBody ServiceRequestDTO serviceRequestDTO
    );

}