# src/main/java/org/example/cuidadodemascotas/servicemicroservice/apis/controller/ServiceTypeApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.example.cuidadodemascotas.servicemicroservice.apis.controller;

import org.example.cuidadodemascotas.servicemicroservice.apis.dto.ServiceTypePageResponse;
import org.example.cuidadodemascotas.servicemicroservice.apis.dto.ServiceTypeRequestDTO;
import org.example.cuidadodemascotas.servicemicroservice.apis.dto.ServiceTypeResponseDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-07T01:35:32.143539700-03:00[America/Asuncion]")
@Validated
@Tag(name = "ServiceType", description = "the ServiceType API")
public interface ServiceTypeApi {

    /**
     * POST /service-types : Crear un nuevo tipo de servicio
     *
     * @param serviceTypeRequestDTO  (required)
     * @return Tipo de servicio creado (status code 201)
     */
    @Operation(
        operationId = "createServiceType",
        summary = "Crear un nuevo tipo de servicio",
        tags = { "ServiceType" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Tipo de servicio creado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ServiceTypeResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/service-types",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ServiceTypeResponseDTO> createServiceType(
        @Parameter(name = "ServiceTypeRequestDTO", description = "", required = true) @Valid @RequestBody ServiceTypeRequestDTO serviceTypeRequestDTO
    );


    /**
     * DELETE /service-types/{id} : Eliminar tipo de servicio
     *
     * @param id  (required)
     * @return Tipo de servicio eliminado (status code 204)
     */
    @Operation(
        operationId = "deleteServiceType",
        summary = "Eliminar tipo de servicio",
        tags = { "ServiceType" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Tipo de servicio eliminado")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/service-types/{id}"
    )
    
    ResponseEntity<Void> deleteServiceType(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    /**
     * GET /service-types/{id} : Obtener tipo de servicio por ID
     *
     * @param id  (required)
     * @return Tipo de servicio obtenido (status code 200)
     */
    @Operation(
        operationId = "getServiceTypeById",
        summary = "Obtener tipo de servicio por ID",
        tags = { "ServiceType" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Tipo de servicio obtenido", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ServiceTypeResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/service-types/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<ServiceTypeResponseDTO> getServiceTypeById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    /**
     * GET /service-types : Obtener todos los tipos de servicio con paginación
     *
     * @param page  (optional, default to 0)
     * @param size  (optional, default to 20)
     * @return Tipos de servicio obtenidos exitosamente (status code 200)
     */
    @Operation(
        operationId = "getServiceTypes",
        summary = "Obtener todos los tipos de servicio con paginación",
        tags = { "ServiceType" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Tipos de servicio obtenidos exitosamente", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ServiceTypePageResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/service-types",
        produces = { "application/json" }
    )
    
    ResponseEntity<ServiceTypePageResponse> getServiceTypes(
        @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
        @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Integer size
    );


    /**
     * PUT /service-types/{id} : Actualizar tipo de servicio
     *
     * @param id  (required)
     * @param serviceTypeRequestDTO  (required)
     * @return Tipo de servicio actualizado (status code 200)
     */
    @Operation(
        operationId = "updateServiceType",
        summary = "Actualizar tipo de servicio",
        tags = { "ServiceType" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Tipo de servicio actualizado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ServiceTypeResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/service-types/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ServiceTypeResponseDTO> updateServiceType(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "ServiceTypeRequestDTO", description = "", required = true) @Valid @RequestBody ServiceTypeRequestDTO serviceTypeRequestDTO
    );

}