# src/main/resources/config-openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Service Microservice API
  description: Microservicio para gestión de servicios y tipos de servicios
  version: 1.0.0

servers:
  - url: http://localhost:8083
    description: Service Microservice

paths:
  # ==================== SERVICE TYPES ====================
  /service-types:
    get:
      tags:
        - ServiceType
      summary: Obtener todos los tipos de servicio con paginación
      operationId: getServiceTypes
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Tipos de servicio obtenidos exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceTypePageResponse'

    post:
      tags:
        - ServiceType
      summary: Crear un nuevo tipo de servicio
      operationId: createServiceType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceTypeRequestDTO'
      responses:
        '201':
          description: Tipo de servicio creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceTypeResponseDTO'

  /service-types/{id}:
    get:
      tags:
        - ServiceType
      summary: Obtener tipo de servicio por ID
      operationId: getServiceTypeById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Tipo de servicio obtenido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceTypeResponseDTO'

    put:
      tags:
        - ServiceType
      summary: Actualizar tipo de servicio
      operationId: updateServiceType
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceTypeRequestDTO'
      responses:
        '200':
          description: Tipo de servicio actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceTypeResponseDTO'

    delete:
      tags:
        - ServiceType
      summary: Eliminar tipo de servicio
      operationId: deleteServiceType
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Tipo de servicio eliminado

  # ==================== SERVICES ====================
  /services:
    get:
      tags:
        - Service
      summary: Obtener servicios con filtros opcionales
      operationId: getServices
      parameters:
        - name: carerId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: serviceTypeId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: minPrice
          in: query
          required: false
          schema:
            type: number
            format: double
        - name: maxPrice
          in: query
          required: false
          schema:
            type: number
            format: double
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Servicios obtenidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServicePageResponse'

    post:
      tags:
        - Service
      summary: Crear un nuevo servicio
      operationId: createService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceRequestDTO'
      responses:
        '201':
          description: Servicio creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceResponseDTO'

  /services/search:
    get:
      tags:
        - Service
      summary: Buscar servicios por descripción
      operationId: searchServicesByText
      parameters:
        - name: text
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Servicios encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServicePageResponse'

  /services/{id}:
    get:
      tags:
        - Service
      summary: Obtener servicio por ID
      operationId: getServiceById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Servicio obtenido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceResponseDTO'

    put:
      tags:
        - Service
      summary: Actualizar servicio
      operationId: updateService
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceRequestDTO'
      responses:
        '200':
          description: Servicio actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceResponseDTO'

    delete:
      tags:
        - Service
      summary: Eliminar servicio (soft delete)
      operationId: deleteService
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Servicio eliminado

components:
  schemas:
    # DTOs de ServiceType
    ServiceTypeRequestDTO:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 100

    ServiceTypeResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string

    ServiceTypePageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ServiceTypeResponseDTO'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer

    # DTOs de Service
    ServiceRequestDTO:
      type: object
      required:
        - carerId
        - serviceTypeId
        - price
      properties:
        carerId:
          type: integer
          format: int64
        serviceTypeId:
          type: integer
          format: int64
        description:
          type: string
        price:
          type: number
          format: double
          minimum: 0

    ServiceResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        carerId:
          type: integer
          format: int64
        serviceTypeId:
          type: integer
          format: int64
        description:
          type: string
        price:
          type: number
          format: double
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        active:
          type: boolean

    ServicePageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ServiceResponseDTO'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer